Feat/add discount

+++ backend/src/clients/database/models/user.py
@@ -13,5 +13,5 @@ class User(Base):
     username: Mapped[str] = mapped_column(nullable=True)
     phone_number: Mapped[str] = mapped_column(nullable=True)
     language_code: Mapped[str] = mapped_column(nullable=True)
-    coins: Mapped[int] = mapped_column(nullable=True)
+    coins: Mapped[float] = mapped_column(nullable=True)
     is_admin: Mapped[bool] = mapped_column(nullable=True)
